Unproven work uploaded
This write intends to used linear algebraic techniques to map a string of binary digits to its representation in base-10 system. This was done as an alternative for lexicographical orderings, however I believe that this is just another to explain lexicographic order for binary strings.

The paper will be edited in OverLeaf so include this and to write an inductive proof.